requirements.txt
opencv-python
flask
gpiozero
rpi.gpio
pyserial
